@@ -3,7 +3,7 @@ package uk.dupplaw.gitlab.wallboard.config
 import io.smallrye.config.ConfigMapping
 import org.eclipse.microprofile.config.inject.ConfigProperty
 import java.util.*
-import javax.enterprise.context.ApplicationScoped
+import jakarta.enterprise.context.ApplicationScoped
 
 @ApplicationScoped
 data class SCMServiceConfiguration(
