@@ -1,7 +1,7 @@
 package uk.dupplaw.gitlab.wallboard.config
 
 import org.eclipse.microprofile.config.inject.ConfigProperty
-import javax.enterprise.context.ApplicationScoped
+import jakarta.enterprise.context.ApplicationScoped
 
 @ApplicationScoped
 data class SystemConfiguration(
