@@ -1,13 +1,13 @@
 package uk.dupplaw.gitlab.wallboard.endpoints
 
 import java.util.regex.Pattern
-import javax.servlet.FilterChain
-import javax.servlet.ServletRequest
-import javax.servlet.ServletResponse
-import javax.servlet.annotation.WebFilter
-import javax.servlet.http.HttpFilter
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.FilterChain
+import jakarta.servlet.ServletRequest
+import jakarta.servlet.ServletResponse
+import jakarta.servlet.annotation.WebFilter
+import jakarta.servlet.http.HttpFilter
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 private val FILE_NAME_PATTERN = Pattern.compile(".*[.][a-zA-Z\\d]+")
 
