@@ -3,10 +3,10 @@ package uk.dupplaw.gitlab.wallboard.endpoints
 import uk.dupplaw.gitlab.wallboard.config.SystemConfiguration
 import uk.dupplaw.gitlab.wallboard.domain.SystemInfo
 import uk.dupplaw.gitlab.wallboard.domain.Version
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.Produces
-import javax.ws.rs.core.MediaType
+import jakarta.ws.rs.GET
+import jakarta.ws.rs.Path
+import jakarta.ws.rs.Produces
+import jakarta.ws.rs.core.MediaType
 
 @Path("/api/system")
 class SystemInfoController(
