@@ -8,7 +8,11 @@ import uk.dupplaw.gitlab.wallboard.domain.BuildService
 import uk.dupplaw.gitlab.wallboard.domain.BuildStatus
 import uk.dupplaw.gitlab.wallboard.domain.Project
 import java.net.URL
-import javax.enterprise.context.ApplicationScoped
+import jakarta.enterprise.context.ApplicationScoped
+
+private const val PROJECT_ID_TAG = ":projectId"
+
+private const val PRIVATE_TOKEN_HEADER = "Private-Token"
 
 @ApplicationScoped
 class GitLabCIBuildService(
@@ -17,10 +21,9 @@ class GitLabCIBuildService(
     private val logger = KotlinLogging.logger {}
 
     private val baseUrl = "/api/v4"
-    private val pipelinesUrl =
-            "$baseUrl/projects/:projectId/pipelines?simple=true&per_page=1&order_by=id&sort=desc&ref=:ref"
-    private val pipelineUrl = "$baseUrl/projects/:projectId/pipelines/:pipelineId?simple=true"
-    private val jobUrl = "$baseUrl/projects/:projectId/pipelines/:pipelineId/jobs?per_page=3&scope=:scope"
+    private val pipelinesUrl = "$baseUrl/projects/$PROJECT_ID_TAG/pipelines?simple=true&per_page=1&order_by=id&sort=desc&ref=:ref"
+    private val pipelineUrl = "$baseUrl/projects/$PROJECT_ID_TAG/pipelines/:pipelineId?simple=true"
+    private val jobUrl = "$baseUrl/projects/$PROJECT_ID_TAG/pipelines/:pipelineId/jobs?per_page=3&scope=:scope"
 
     override fun retrieveBuildInformation(project: Project) =
             getLatestBuild(project.id)?.let {
@@ -29,7 +32,7 @@ class GitLabCIBuildService(
 
     private fun getLatestBuild(projectId: Long): Long? {
         val projectUrl = "https://${gitLabCIBuildServiceConfiguration.host}$pipelinesUrl"
-                .replace(":projectId", projectId.toString())
+                .replace(PROJECT_ID_TAG, projectId.toString())
                 .replace(
                         ":ref",
                         gitLabCIBuildServiceConfiguration.overriddenRefs.refs()[projectId]
@@ -40,7 +43,7 @@ class GitLabCIBuildService(
         return URL(projectUrl).openConnection().apply {
             readTimeout = 800
             connectTimeout = 200
-            setRequestProperty("Private-Token", gitLabCIBuildServiceConfiguration.token)
+            setRequestProperty(PRIVATE_TOKEN_HEADER, gitLabCIBuildServiceConfiguration.token)
         }.getInputStream().use { ins ->
             ObjectMapper().readTree(ins).firstOrNull()?.get("id")?.asLong()
         }
@@ -48,18 +51,18 @@ class GitLabCIBuildService(
 
     private fun getBuildInfo(projectId: Long, buildId: Long): Build {
         val buildUrl = "https://${gitLabCIBuildServiceConfiguration.host}$pipelineUrl"
-                .replace(":projectId", projectId.toString())
+                .replace(PROJECT_ID_TAG, projectId.toString())
                 .replace(":pipelineId", buildId.toString())
 
         logger.trace { "Getting build info: $buildUrl" }
         return URL(buildUrl).openConnection().apply {
             readTimeout = 800
             connectTimeout = 200
-            setRequestProperty("Private-Token", gitLabCIBuildServiceConfiguration.token)
+            setRequestProperty(PRIVATE_TOKEN_HEADER, gitLabCIBuildServiceConfiguration.token)
         }.getInputStream().use { ins ->
             ObjectMapper().readTree(ins).let { node ->
-                val id = node.get("id").asLong()
-                val actualStatus = node.get("status").asText()
+                val id = node["id"].asLong()
+                val actualStatus = node["status"].asText()
                 val status = when (actualStatus) {
                     "success" -> BuildStatus.SUCCESS
                     "failed" -> BuildStatus.FAIL
@@ -67,9 +70,9 @@ class GitLabCIBuildService(
                     "created", "waiting_for_resources", "preparing", "pending", "running" -> BuildStatus.RUNNING
                     else -> BuildStatus.UNKNOWN
                 }
-                val lastBuild = node.get("updated_at").asText()
-                val webUrl = node.get("web_url").asText()
-                val user = node.get("user").get("name").asText()
+                val lastBuild = node["updated_at"].asText()
+                val webUrl = node["web_url"].asText()
+                val user = node["user"]["name"].asText()
                 val scope = when (status) {
                     BuildStatus.FAIL, BuildStatus.WARNING, BuildStatus.RUNNING -> actualStatus
                     else -> null
@@ -92,7 +95,7 @@ class GitLabCIBuildService(
     fun getJobInfo(projectId: Long, pipelineId: Long, scope: String): List<String> {
         try {
             val jobUrl = "https://${gitLabCIBuildServiceConfiguration.host}$jobUrl"
-                    .replace(":projectId", projectId.toString())
+                    .replace(PROJECT_ID_TAG, projectId.toString())
                     .replace(":pipelineId", pipelineId.toString())
                     .replace(":scope", scope)
 
@@ -101,9 +104,9 @@ class GitLabCIBuildService(
             return URL(jobUrl).openConnection().apply {
                 readTimeout = 800
                 connectTimeout = 200
-                setRequestProperty("Private-Token", gitLabCIBuildServiceConfiguration.token)
+                setRequestProperty(PRIVATE_TOKEN_HEADER, gitLabCIBuildServiceConfiguration.token)
             }.getInputStream().use { ins ->
-                ObjectMapper().readTree(ins).map { it.get("name").asText() }
+                ObjectMapper().readTree(ins).map { it["name"].asText() }
             }
         } catch (e: Exception) {
             logger.warn { "Error getting job information" }
